Estimate location

Calculate point estimates based on antenna bearing and signal strength.

locate(
  data = NULL,
  ID = "tagDeployID",
  ts = "ts",
  sig = "sig",
  aLon = "recvDeployLon",
  aLat = "recvDeployLat",
  aBearing = "antBearing",
  det_range = 12,
  dtime = 2
)

Arguments

data

A data.frame containing the telemetry data.

ID

Unique identifier for individuals or tag deployments.

ts

Timestamp column.

sig

Signal strength column.

aLon

Antenna longitude column.

aLat

Antenna latitude column.

aBearing

Antenna bearing column.

det_range

Assumed maximum detection range of antennas in kilometres.

dtime

Time interval in minutes for which point estimates are to be calculated.

Value

Returns a data.frame containing estimated coordinates and measurement errors for each time interval together with the proportions of time intervals w.

Details

Estimate location for each detection: half of detection range along the directional beam.

Derive oscillating measurement error.

Weighted means (by signal strength) for each time interval.